+++ src/main/java/com/umc/networkingService/domain/member/client/GoogleMemberClient.java
@@ -24,7 +24,7 @@ public class GoogleMemberClient {
                 .block();
 
         if(response == null)
-            throw new RestApiException(ErrorCode._INTERNAL_SERVER_ERROR);
+            throw new RestApiException(ErrorCode.FAILED_SOCIAL_LOGIN);
 
         return response.getSub();
     }
