+++ src/main/java/com/umc/networkingService/domain/member/client/KakaoMemberClient.java
@@ -26,7 +26,7 @@ public class KakaoMemberClient {
                 .block();
 
         if(response == null)
-            throw new RestApiException(ErrorCode._INTERNAL_SERVER_ERROR);
+            throw new RestApiException(ErrorCode.FAILED_SOCIAL_LOGIN);
 
         return response.getId();
     }
