+++ src/main/java/com/umc/networkingService/domain/member/controller/AuthController.java
@@ -4,10 +4,14 @@ import com.umc.networkingService.config.security.auth.CurrentMember;
 import com.umc.networkingService.domain.member.dto.request.MemberSignUpRequest;
 import com.umc.networkingService.domain.member.dto.response.MemberGenerateNewAccessTokenResponse;
 import com.umc.networkingService.domain.member.dto.response.MemberIdResponse;
+import com.umc.networkingService.domain.member.dto.response.MemberLoginResponse;
 import com.umc.networkingService.domain.member.entity.Member;
+import com.umc.networkingService.domain.member.entity.SocialType;
 import com.umc.networkingService.domain.member.service.AuthService;
 import com.umc.networkingService.global.common.base.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -21,6 +25,19 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/members")
 public class AuthController {
     private final AuthService authService;
+
+    @Operation(summary = "소셜 로그인", description = "네이버, 카카오, 구글, 애플 로그인")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "로그인 성공"),
+            @ApiResponse(responseCode = "AUTH007", description = "외부 소셜 서버와의 통신 에러" , content =
+            @Content(schema = @Schema(implementation = BaseResponse.class)))
+    })
+    @PostMapping("/login")
+    public BaseResponse<MemberLoginResponse> socialLogin(@RequestParam(value = "accessToken") String accessToken,
+                                                         @RequestParam(value = "socialType") SocialType socialType) {
+        return BaseResponse.onSuccess(authService.socialLogin(accessToken, socialType));
+
+    }
 
     @Operation(summary = "회원가입 API", description = "최초 멤버 정보를 등록하는 API입니다.")
     @ApiResponses( value = {
