+++ src/main/java/com/umc/networkingService/domain/member/controller/StaffMemberController.java
@@ -1,20 +1,5 @@
 package com.umc.networkingService.domain.member.controller;
 
-import com.umc.networkingService.config.security.auth.CurrentMember;
-import com.umc.networkingService.domain.member.dto.request.MemberUpdateProfileRequest;
-import com.umc.networkingService.domain.member.dto.response.MemberIdResponse;
-import com.umc.networkingService.domain.member.entity.Member;
-import com.umc.networkingService.domain.member.service.MemberService;
-import com.umc.networkingService.global.common.base.BaseResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.UUID;
-
 // TODO: security 문제 해결 후 개발
 //@Tag(name = "멤버 API", description = "운영진용 멤버 관련 API")
 //@RestController
