+++ src/main/java/com/umc/networkingService/domain/member/dto/request/MemberSignUpRequest.java
@@ -1,8 +1,7 @@
 package com.umc.networkingService.domain.member.dto.request;
 
-import com.umc.networkingService.global.common.enums.Part;
-import com.umc.networkingService.global.common.enums.Semester;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -19,6 +18,7 @@ public class MemberSignUpRequest {
     private String name;
     @NotBlank(message = "닉네임은 필수 입력값입니다.")
     private String nickname;
+    @Size(min = 1, message = "최소 하나의 기수별 파트가 필요합니다.")
     private List<SemesterPartInfo> semesterParts;
     private String universityName;
     private List<String> campusPositions;
