+++ src/main/java/com/umc/networkingService/domain/member/dto/response/MemberLoginResponse.java
@@ -1,6 +1,5 @@
 package com.umc.networkingService.domain.member.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
