+++ src/main/java/com/umc/networkingService/domain/member/repository/MemberRepository.java
@@ -1,11 +1,8 @@
 package com.umc.networkingService.domain.member.repository;
 
 import com.umc.networkingService.domain.member.entity.Member;
-import com.umc.networkingService.domain.university.entity.University;
 import com.umc.networkingService.domain.member.entity.SocialType;
-import java.util.Optional;
-import java.util.UUID;
-
+import com.umc.networkingService.domain.university.entity.University;
 import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
