+++ src/main/java/com/umc/networkingService/domain/member/service/MemberService.java
@@ -6,13 +6,10 @@ import com.umc.networkingService.domain.member.dto.response.*;
 import com.umc.networkingService.domain.member.entity.Member;
 import com.umc.networkingService.global.common.base.EntityLoader;
 import org.springframework.web.multipart.MultipartFile;
-import com.umc.networkingService.domain.member.dto.response.MemberLoginResponse;
-import com.umc.networkingService.domain.member.entity.SocialType;
 
 import java.util.UUID;
 
 public interface MemberService extends EntityLoader<Member, UUID> {
-    MemberLoginResponse socialLogin(final String accessToken, SocialType socialType);
     MemberIdResponse updateMyProfile(Member member, MultipartFile profileImage, MemberUpdateMyProfileRequest request);
     MemberIdResponse updateProfile(Member member, UUID memberId, MemberUpdateProfileRequest request);
     MemberInquiryProfileResponse inquiryProfile(Member member, UUID memberId);
