+++ src/main/java/com/umc/networkingService/domain/member/service/MemberServiceImpl.java
@@ -1,10 +1,8 @@
 package com.umc.networkingService.domain.member.service;
 
 
-import com.umc.networkingService.config.security.jwt.JwtTokenProvider;
-import com.umc.networkingService.config.security.jwt.TokenInfo;
 import com.umc.networkingService.domain.friend.service.FriendService;
-import com.umc.networkingService.domain.member.client.*;
+import com.umc.networkingService.domain.member.client.GithubMemberClient;
 import com.umc.networkingService.domain.member.dto.request.MemberUpdateMyProfileRequest;
 import com.umc.networkingService.domain.member.dto.request.MemberUpdateProfileRequest;
 import com.umc.networkingService.domain.member.dto.response.*;
@@ -32,17 +30,10 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class MemberServiceImpl implements MemberService{
 
-    private final KakaoMemberClient kakaoMemberClient;
-    private final GoogleMemberClient googleMemberClient;
-    private final NaverMemberClient naverMemberClient;
-    private final AppleMemberClient appleMemberClient;
+
     private final MemberMapper memberMapper;
-
-    private final MemberPointRepository memberPointRepository;
     private final MemberRepository memberRepository;
-
-    private final JwtTokenProvider jwtTokenProvider;
-    private final RefreshTokenService refreshTokenService;
+    private final MemberPointRepository memberPointRepository;
 
     private final SemesterPartService semesterPartService;
     private final MemberPositionService memberPositionService;
@@ -50,98 +41,6 @@ public class MemberServiceImpl implements MemberService{
 
     private final S3FileComponent s3FileComponent;
     private final GithubMemberClient githubMemberClient;
-
-    @Override
-    public MemberLoginResponse socialLogin(String accessToken, SocialType socialType){
-        // 로그인 구분
-        if(socialType.equals(SocialType.KAKAO))
-            return loginByKakao(accessToken);
-
-        if(socialType.equals(SocialType.GOOGLE))
-            return loginByGoogle(accessToken);
-
-        if(socialType.equals(SocialType.NAVER))
-            return loginByNaver(accessToken);
-
-        if(socialType.equals(SocialType.APPLE))
-            return loginByApple(accessToken);
-
-        return null;
-    }
-
-    private MemberLoginResponse loginByApple(final String accessToken){
-        // apple 서버와 통신해서 유저 고유값(clientId) 받기
-        String clientId = appleMemberClient.getappleClientID(accessToken);
-        //존재 여부 파악
-        Optional<Member> getMember = memberRepository.findByClientIdAndSocialType(clientId, SocialType.APPLE);
-
-        //1. 없으면 : Member 객체 생성하고 DB 저장
-        if(getMember.isEmpty()){
-            return saveNewMember(clientId, SocialType.APPLE);
-        }
-        // 2. 있으면 : 새로운 토큰 반환
-        return getNewToken(getMember.get());
-    }
-
-    private MemberLoginResponse loginByKakao(final String accessToken){
-        // kakao 서버와 통신해서 유저 고유값(clientId) 받기
-        String clientId = kakaoMemberClient.getkakaoClientID(accessToken);
-        // 존재 여부 파악
-        Optional<Member> getMember = memberRepository.findByClientIdAndSocialType(clientId, SocialType.KAKAO);
-
-        // 1. 없으면 : Member 객체 생성하고 DB 저장
-        if(getMember.isEmpty()) {
-            return saveNewMember(clientId, SocialType.KAKAO);
-        }
-        // 2. 있으면 : 새로운 토큰 반환
-        return getNewToken(getMember.get());
-    }
-
-    private MemberLoginResponse loginByNaver(final String accessToken){
-        // naver 서버와 통신해서 유저 고유값(clientId) 받기
-        String clientId = naverMemberClient.getnaverClientID(accessToken);
-        // 존재 여부 파악
-        Optional<Member> getMember = memberRepository.findByClientIdAndSocialType(clientId,SocialType.NAVER);
-
-        // 1. 없으면 (처음 로그인 하는 경우)
-        if(getMember.isEmpty()) {
-           return saveNewMember(clientId,SocialType.NAVER);
-        }
-        // 2. 있으면 (이미 로그인 했던 적이 있는 경우)
-        return getNewToken(getMember.get());
-    }
-
-    private MemberLoginResponse loginByGoogle(final String accessToken){
-        // google 서버와 통신해서 유저 고유값(clientId) 받기
-        String clientId = googleMemberClient.getgoogleClientID(accessToken);
-        // 존재 여부 파악
-        Optional<Member> getMember = memberRepository.findByClientIdAndSocialType(clientId, SocialType.GOOGLE);
-
-        // 1. 없으면 : Member 객체 생성하고 DB 저장
-        if(getMember.isEmpty()){
-            return saveNewMember(clientId, SocialType.GOOGLE);
-        }
-        // 2. 있으면 : 새로운 토큰 반환
-        return getNewToken(getMember.get());
-    }
-
-    private MemberLoginResponse saveNewMember(String clientId, SocialType socialType) {
-        Member member = memberMapper.toMember(clientId, socialType);
-        Member newMember =  memberRepository.save(member);
-
-        return getNewToken(newMember);
-    }
-
-    private MemberLoginResponse getNewToken(Member member) {
-        // jwt 토큰 생성
-        TokenInfo tokenInfo = jwtTokenProvider.generateToken(member.getId());
-        // refreshToken 디비에 저장
-        refreshTokenService.saveTokenInfo(tokenInfo.getRefreshToken(), member.getId());
-
-        return memberMapper.toLoginMember(member, tokenInfo);
-    }
-
-
 
     // 나의 프로필 업데이트 함수
     @Override
@@ -194,7 +93,9 @@ public class MemberServiceImpl implements MemberService{
 
     // 프로필 조회 함수
     @Override
-    public MemberInquiryProfileResponse inquiryProfile(Member member, UUID memberId) {
+    public MemberInquiryProfileResponse inquiryProfile(Member loginMember, UUID memberId) {
+
+        Member member = loadEntity(loginMember.getId());
         // 본인 프로필 조회인 경우
         if (memberId == null) {
             return memberMapper.toInquiryProfileResponse(member, MemberRelation.MINE);
@@ -213,7 +114,10 @@ public class MemberServiceImpl implements MemberService{
 
     // 포인트 관련 정보 조회
     @Override
-    public MemberInquiryInfoWithPointResponse inquiryInfoWithPoint(Member member) {
+    public MemberInquiryInfoWithPointResponse inquiryInfoWithPoint(Member loginMember) {
+
+        Member member = loadEntity(loginMember.getId());
+
         // 소속 대학교 찾기
         University university = Optional.ofNullable(member.getUniversity())
                 .orElseThrow(() -> new RestApiException(ErrorCode.EMPTY_MEMBER_UNIVERSITY));
@@ -247,12 +151,14 @@ public class MemberServiceImpl implements MemberService{
 
         String gitNickName = member.getGitNickname();
         if (gitNickName == null)
-            throw new RestApiException(ErrorCode.UNAUTHENTICATION_GITHUB);
+            throw new RestApiException(ErrorCode.UNAUTHENTICATED_GITHUB);
         return new MemberInquiryGithubResponse("https://ghchart.rshah.org/2965FF/" + gitNickName);
     }
 
     @Override
-    public MemberInquiryPointsResponse inquiryMemberPoints(Member member) {
+    public MemberInquiryPointsResponse inquiryMemberPoints(Member loginMember) {
+        Member member = loadEntity(loginMember.getId());
+
         Page<MemberPoint> usedPointsPage = memberPointRepository.
                 findAllByMemberOrderByCreatedAtDesc(member, PageRequest.of(0, 2));
         List<MemberInquiryPointsResponse.UsedHistory> usedHistories = usedPointsPage.stream()
