+++ src/main/java/com/umc/networkingService/domain/member/service/SemesterPartService.java
@@ -1,6 +1,5 @@
 package com.umc.networkingService.domain.member.service;
 
-import com.umc.networkingService.domain.member.dto.request.MemberSignUpRequest;
 import com.umc.networkingService.domain.member.dto.request.SemesterPartInfo;
 import com.umc.networkingService.domain.member.entity.Member;
 import com.umc.networkingService.domain.member.entity.SemesterPart;
