+++ src/test/java/com/umc/networkingService/domain/member/service/AuthServiceIntegrationTest.java
@@ -1,27 +1,128 @@
 package com.umc.networkingService.domain.member.service;
 
+import com.umc.networkingService.domain.member.client.GoogleMemberClient;
+import com.umc.networkingService.domain.member.client.KakaoMemberClient;
+import com.umc.networkingService.domain.member.client.NaverMemberClient;
 import com.umc.networkingService.domain.member.dto.request.MemberSignUpRequest;
 import com.umc.networkingService.domain.member.dto.request.SemesterPartInfo;
 import com.umc.networkingService.domain.member.dto.response.MemberGenerateNewAccessTokenResponse;
+import com.umc.networkingService.domain.member.dto.response.MemberLoginResponse;
 import com.umc.networkingService.domain.member.entity.Member;
+import com.umc.networkingService.domain.member.entity.RefreshToken;
+import com.umc.networkingService.domain.member.entity.SocialType;
 import com.umc.networkingService.global.common.enums.Part;
+import com.umc.networkingService.global.common.enums.Role;
 import com.umc.networkingService.global.common.enums.Semester;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
 
 @DisplayName("Auth 서비스의 ")
 @SpringBootTest
 public class AuthServiceIntegrationTest extends MemberServiceTestConfig {
 
     @Autowired AuthService authService;
+
+    @MockBean private KakaoMemberClient kakaoMemberClient;
+    @MockBean private GoogleMemberClient googleMemberClient;
+    @MockBean private NaverMemberClient naverMemberClient;
+
+    @Test
+    @DisplayName("카카오 로그인 테스트")
+    @Transactional
+    public void kakaoLoginTest() {
+        // given
+        String accessToken = "ya29.a0AfB_byD6";
+        String clientId = "abcdefg";
+
+        given(kakaoMemberClient.getkakaoClientID(any())).willReturn(clientId);
+
+        //when
+        MemberLoginResponse response = authService.socialLogin(accessToken, SocialType.KAKAO);
+
+        //then
+        // 멤버 저장 상태 테스트
+        Optional<Member> optionalMember = memberRepository.findById(response.getMemberId());
+        assertTrue(optionalMember.isPresent());
+        Member savedMember = optionalMember.get();
+
+        assertEquals(clientId, savedMember.getClientId());
+        assertEquals(Role.MEMBER, savedMember.getRole());
+        assertEquals(SocialType.KAKAO, savedMember.getSocialType());
+
+        // 리프레시 토큰 저장 상태 테스트
+        RefreshToken refreshToken = refreshTokenService.findByMemberId(savedMember.getId())
+                .orElseThrow();
+        assertEquals(response.getRefreshToken(), refreshToken.getRefreshToken());
+    }
+
+    @Test
+    @DisplayName("구글 로그인 테스트")
+    @Transactional
+    public void googleLoginTest() {
+        // given
+        String accessToken = "ya29.a0AfB_byD6";
+        String sub = "abcdefg";
+
+        given(googleMemberClient.getgoogleClientID(any())).willReturn(sub);
+
+        //when
+        MemberLoginResponse response = authService.socialLogin(accessToken, SocialType.GOOGLE);
+
+        //then
+        // 멤버 저장 상태 테스트
+        Optional<Member> optionalMember = memberRepository.findById(response.getMemberId());
+        assertTrue(optionalMember.isPresent());
+        Member savedMember = optionalMember.get();
+
+        assertEquals(sub, savedMember.getClientId());
+        assertEquals(Role.MEMBER, savedMember.getRole());
+        assertEquals(SocialType.GOOGLE, savedMember.getSocialType());
+
+        // 리프레시 토큰 저장 상태 테스트
+        RefreshToken refreshToken = refreshTokenService.findByMemberId(savedMember.getId())
+                .orElseThrow();
+        assertEquals(response.getRefreshToken(), refreshToken.getRefreshToken());
+    }
+
+    @Test
+    @DisplayName("네이버 로그인 테스트")
+    @Transactional
+    public void naverLoginTest() {
+        // given
+        String accessToken = "ya29.a0AfB_byD6";
+        String clientId = "abcdefg";
+
+        given(naverMemberClient.getnaverClientID(any())).willReturn(clientId);
+
+        //when
+        MemberLoginResponse response = authService.socialLogin(accessToken, SocialType.NAVER);
+
+        //then
+        // 멤버 저장 상태 테스트
+        Optional<Member> optionalMember = memberRepository.findById(response.getMemberId());
+        assertTrue(optionalMember.isPresent());
+        Member savedMember = optionalMember.get();
+
+        assertEquals(clientId, savedMember.getClientId());
+        assertEquals(Role.MEMBER, savedMember.getRole());
+        assertEquals(SocialType.NAVER, savedMember.getSocialType());
+
+        // 리프레시 토큰 저장 상태 테스트
+        RefreshToken refreshToken = refreshTokenService.findByMemberId(savedMember.getId())
+                .orElseThrow();
+        assertEquals(response.getRefreshToken(), refreshToken.getRefreshToken());
+    }
 
 
     @Test
